# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Voice Agent Bridge API
  version: 0.1.0
  description: |
    HTTP API for Voice Agent Bridge (VAB).

    **Security**
    - `UserBearer` — JWT from `/v1/auth/register` or `/v1/auth/login`
    - `AgentKey` — header `X-Agent-Key` (per-agent key from create/pairing claim)

    **Push semantics**
    - `POST /v1/sessions/{id}/progress` (**update_progress**) NEVER pushes
    - `POST /v1/sessions/{id}/events` (**report_event**) MAY push for `needs_user`,
      or `succeeded`/`failed` with `actions` / `force_push`
    - Agent decides `needs_user`; the bridge does not guess

    **TTL**
    - Session default max 24h (`86400`)
    - Destructive confirm window max 30m (`1800`)
servers:
  - url: http://127.0.0.1:8787
    description: Local MVP
...
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Register a user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "409":
          $ref: "#/components/responses/Error"
  /v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Login and obtain JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "401":
          $ref: "#/components/responses/Error"
  /v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refresh
      summary: Rotate a refresh token and issue a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Rotated token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokenResponse" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke a refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Revoked
  /v1/agents:
    get:
      tags: [agents]
      operationId: listAgents
      summary: List agents for the authenticated user
      security:
        - UserBearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [agents]
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Agent"
    post:
      tags: [agents]
      operationId: createAgent
      summary: Create an agent; returns api_key once
      security:
        - UserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              additionalProperties: false
              properties:
                label: { type: string, minLength: 1 }
                host_label: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentCreated"
  /v1/agents/{agent_id}/rotate-key:
    post:
      tags: [agents]
      operationId: rotateAgentKey
      summary: Revoke the current agent key and issue a replacement
      security:
        - UserBearer: []
      parameters:
        - name: agent_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Replacement key, shown only once
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentCreated" }
  /v1/pairing/code:
    post:
      tags: [pairing]
      operationId: createPairingCode
      summary: Create a short-lived pairing code
      security:
        - UserBearer: []
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [code, expires_at]
                properties:
                  code: { type: string }
                  expires_at: { type: string }
  /v1/pairing/claim:
    post:
      tags: [pairing]
      operationId: claimPairingCode
      summary: Claim pairing code; creates agent + api_key (public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, label]
              additionalProperties: false
              properties:
                code: { type: string }
                label: { type: string, minLength: 1 }
                host_label: { type: string }
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentCreated"
        "400":
          $ref: "#/components/responses/Error"
  /v1/skills:
    get:
      tags: [skills]
      operationId: listSkills
      summary: List skills
      security:
        - UserBearer: []
        - AgentKey: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [skills]
                properties:
                  skills:
                    type: array
                    items:
                      $ref: "#/components/schemas/Skill"
    post:
      tags: [skills]
      operationId: upsertSkill
      summary: Create or upsert a skill
      security:
        - UserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Skill"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /v1/sessions:
    post:
      tags: [sessions]
      operationId: createOrResumeSession
      summary: Create or resume a session (agent)
      description: Session is created when the agent calls this with `skill_id`.
      security:
        - AgentKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_id]
              additionalProperties: false
              properties:
                skill_id: { type: string }
                session_id: { type: string }
                idempotency_key: { type: string }
                chat_id: { type: string }
                title: { type: string }
                facts:
                  type: object
                  additionalProperties: true
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Created or resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/agents/me/actions/pending:
    get:
      tags: [actions]
      operationId: getAgentPendingActions
      summary: List/claim queued actions for this agent
      security:
        - AgentKey: []
      parameters:
        - $ref: "#/components/parameters/Claim"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingActionsResponse"
  /v1/actions/{action_id}/result:
    post:
      tags: [actions]
      operationId: submitActionResult
      summary: Submit result after executing a claimed action
      security:
        - AgentKey: []
      parameters:
        - name: action_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ok]
              additionalProperties: false
              properties:
                ok: { type: boolean }
                message: { type: string }
                output:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Result accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
  /v1/phone/sessions:
    get:
      tags: [phone]
      operationId: listPhoneSessions
      summary: List non-closed sessions for the phone user
      security:
        - UserBearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Session"
...
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT from /v1/auth/register or /v1/auth/login
    AgentKey:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: Per-agent API key from agent create or pairing claim
  parameters:
...
    Claim:
      name: claim
      in: query
      description: When not "false", claim queued actions (default true)
      schema: { type: boolean, default: true }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    ProgressStatus:
      type: string
      enum: [started, running, blocked, succeeded, failed, cancelled]
      description: Progress update status — NEVER triggers push
...
    SessionState:
      type: string
      enum:
        - open
        - running
        - needs_user
        - awaiting_confirm
        - queued
        - claimed
        - closed
        - expired
    ActionRisk:
      type: string
      enum: [low, medium, high, destructive]
      description: destructive (or confirm:true) requires second confirm (≤30m)
    ActionStatus:
      type: string
      enum:
        - offered
        - pending_confirm
        - awaiting_confirm
        - queued
        - claimed
        - done
        - failed
        - cancelled
        - expired
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    AuthCredentials:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    AuthTokenResponse:
      type: object
      required: [user_id, token, refresh_token, expires_in]
      properties:
        user_id: { type: string }
        token: { type: string, description: JWT for UserBearer }
        refresh_token: { type: string, description: One-time-rotatable opaque token }
        expires_in: { type: integer, description: Access-token lifetime in seconds }
    Agent:
      type: object
      required: [agent_id, label, created_at]
      properties:
        agent_id: { type: string }
        label: { type: string }
        host_label: { type: ["string", "null"] }
        created_at: { type: string }
        last_seen_at: { type: ["string", "null"] }
    AgentCreated:
      type: object
      required: [agent, api_key]
      properties:
        agent:
          $ref: "#/components/schemas/Agent"
        api_key:
          type: string
          description: Shown only once; use as X-Agent-Key
...
    SkillActionDef:
      type: object
      required: [id, risk, confirm]
      properties:
        id: { type: string }
        risk:
          $ref: "#/components/schemas/ActionRisk"
        confirm: { type: boolean }
        title: { type: string }
    Skill:
      type: object
      required: [skill_id, template, facts_schema, actions, ttl]
      properties:
        skill_id: { type: string, examples: [deploy.result] }
        template:
          type: string
          examples: ["{{service}} 在 {{env}} 部署{{status}}"]
        facts_schema:
          type: array
          items: { type: string }
        actions:
          type: array
          items:
            $ref: "#/components/schemas/SkillActionDef"
        ttl:
          type: object
          required: [default_sec, destructive_sec]
          properties:
            default_sec:
              type: integer
              maximum: 86400
            destructive_sec:
              type: integer
              maximum: 1800
        version: { type: integer }
    Session:
      type: object
      required: [session_id, skill_id, state]
      properties:
        session_id: { type: string }
        user_id: { type: string }
        agent_id: { type: string }
        skill_id: { type: string }
        state:
          $ref: "#/components/schemas/SessionState"
        chat_id: { type: ["string", "null"] }
        title: { type: ["string", "null"] }
        progress_status:
          oneOf:
            - $ref: "#/components/schemas/ProgressStatus"
            - type: "null"
        progress_percent: { type: ["number", "null"], minimum: 0, maximum: 100 }
        progress_message: { type: ["string", "null"] }
        summary_text: { type: ["string", "null"] }
        voice_script:
          type: ["string", "null"]
          description: Summary truncated ~80 chars for voice
        available_actions:
          type: array
          items: { type: string }
        facts:
          type: object
          additionalProperties: true
        expires_at: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
...
    Action:
      type: object
      required: [action_id, session_id, action_key, risk, status]
      properties:
        action_id: { type: string }
        session_id: { type: string }
        action_key: { type: string }
        title: { type: [string, "null"] }
        risk:
          $ref: "#/components/schemas/ActionRisk"
        confirm_required: { type: boolean }
        status:
          $ref: "#/components/schemas/ActionStatus"
        payload:
          type: object
          additionalProperties: true
        expires_at: { type: string }
        claimed_at: { type: ["string", "null"] }
        result:
          type: [object, "null"]
          additionalProperties: true
        cancelled_by_user: { type: boolean }
    PendingActionsResponse:
      type: object
      required: [actions]
      properties:
        actions:
          type: array
          items:
            $ref: "#/components/schemas/Action"
...